+++ src/pizzaShop/Order.java
@@ -2,7 +2,7 @@ package pizzaShop;
 
 public abstract class Order{
 	abstract void TakeOut();
-	abstract void TakeOut(double OrderNumber, double Total);
+	abstract void TakeOut(int OrderNumber, int Total);
 	
 
 }
