+++ src/pizzaShop/ShopInformation.java
@@ -3,5 +3,5 @@ package pizzaShop;
 public interface ShopInformation {
 	String nameOfCustomer();
 	String nameOfEmployee();
-	double orderNumber(double Order, double total );
+	int orderNumber(int Order, int total );
 }
